Plans can be persisted to the database
When a user arrives on the service the first transaction we are offering is to help them complete a Planning phase.

They will be asked a series of questions, all of which will be defined by the Contentful API. Whilst the questions will be sourced remotely, the user will be able to create plans which bring those questions and their answers together. Users should be able to view these and come back to them in future, which makes it stand out as something we want to include as a first-class resource within this service.

Plans will also be available to individual users. In future we can update Plans to include email addresses or any other user identifier. This will enable us to list out all plans that a user is authorised to see.

Catering is the first category we as a team are addressing in private beta. Whilst in public beta we will want to add others such as Energy. Category is the most fundamental piece of information we need so we start with that.