Use a service to get contentful responses
Isolating this into a service allows us to keep services doing 1 thing well as well making it easier to stub out the boundaries in our test suite.

The plan is that the very next slice of this story will focus on replacing the use of a fixture with a real request.