Use a service to create a local copy of a question
We know we want to keep a local copy of the question as prepared by contentful at the time the journey begins. We need some stabilitiy to present questions and return answers whilst using contentful which could change under us.

There is enough leg work needed in this function that extracting out of the controller and into the service feels clean and testable.