Plans can have many associated Questions
Store the attributes we need from Contentful in order to present a question to the user of the service.

We also store the raw JSON alongside it for reference. It includes additional information such as version. Using the version as an example, it is something we may wish to add as a field later on. Storing the JSON from the off could provide the opportunity to do a retrospective migration to populate those fields.

In a content led service it also gives us the team confidence on what data was received from Contentful at the time. This can help us with debugging.

There is no real ongoing cost to keeping this data. If it becomes a problem later or if by the end we realise we don't need it, it's cheap to drop.