Questions can have an Answer
Having a separate model for answers allows us to use clean routing and validations.

We can extend answer behaviour separately from the question. We may need Radio answers and Text answers with different validation as an example. Right now we could have bundled this in as a field on the question and made it work but this way round keeps small objects with single responsibilities.