Planning form is styled with GOV.UK styles
Now we are using a proper form builder I could do iterate the form to using an @answer object. Though it's not used for validation yet it does prepare us for that.

A question object now responds to `radio_options` to assist us with creating plain old ruby objects that the form builder needs. It expects to be able to call `id` and `name` on options passed to it. These expectations may change for different types of form input so "radio" is prefixed to make that easier to extend in future.