Answers validate the presence of a response
The formbuilder gem takes take of the validations that are defined on the model. The only change needed in the form was to add a hidden field for a nil value so that if nothing was selected a value of nil was still passed to the contorller. Without this we get strong parameters errors which complain there is no 'answer' at all.

We make local variables into instance variables since they're needed when re-rendering the questions new page in the case of errors.