Reuse Contentful client stub by moving to a helper
We now want to add more tests that need to go through the same set up for creating a Contentful client before cutting to what they need to test.

To avoid repeating ourselves with lots of client stubbing we create a reusable method.

We leave a single happy path test that continues to assert that the client initialisation is called with the correct values from our environment variables: CONTENTFUL_SPACE and CONTENTFUL_ACCESS_TOKEN. This is important to leave in as it remains the only check to make sure the client init is still tested.