Refactor how the next entry ID is determined by storing it on the plan
Whilst what we had before with using ENV in `GetContentfulQuestion` works OK for getting the starter question, it doesn't help us when we need to determine what the next question from there should be.

Now we're considering that problem we can use the Plan record to store this value. This approach has a couple of advantages:

1. if we need to update an entry_id  for plans that have been started, we can go through all plan records and update the column where we find a match
2. we can use the same mechanism for determining the first question as well as the second and nth question.