Starter test fixture simplified by removing next question
As we add new logic to handle questions that link to other questions we have a test suite that needs to be updated in order to stay green. This change removes 'next' details from the JSON response by immitating another entry that has no linked entry. This allows us to keep the same original fixture for the existing test suite, but not block the new work.