Show an error if the next question isn't in an expected structure
We have seen in the prototype that some pages have been added which are not concerned with asking questions. The example we have is `unmanagedPage` which looks to be responsible for creating a confirmation page.

There is an assumption in the alpha that Contentful will also be responsible for generating confirmation pages. This may still end up being the case but until we have clear user needs to help us decide I have tried to keep the functionality focused on a linear sequence of expected questions. If a question is not found we explode loudly and we take the needs of that new type into account and move forward from there.