Handle the case where an expected Contentful Entry was not found
It seems like the Contentful Ruby gem does some error handling so that when we call `.entry` we receive `nil` as the response. Since retrieving questions (especially the starter question) is a critical part of the service we want to fail loudly.

The user cannot recover as all questions are currently sourced from Contentful. This may change in future with caching as a way to mitigate this.

The team need to know as soon as possible that an entry the service expected has been removed. The simplest way to start is to send a message to Rollbar. From here we can set up integrations for certain alerts to go to the team Slack channel.

The rollbar event includes 2 pieces of technical information that the user cannot see on the error page: entry_id and space_id. We could also display these on the service error page but I'd like to see if we can get the team familiar with Rollbar and avoid exposing more information than we need to.