Rename the service for getting Contentful entries
This was needlessly tied to "Question" by it's previous name. By renaming it consistently use Entry and matching it up with the resource it deals with (Contentful Entries) we hopefully make it easier to understand. We also make it clearer that we expect that it could be reused for getting contentful decisions, which are different types of entries.