Remove unneccessary error message usage
We are no longer using the error message in the view since removing the entry_id from the message. This copy is only used for rollbar so we can simplify this setup.

To pass the entry_id to private methods without having to hand it around as an internal parameter, we move it to being required as part of the constructor. This makes sense as it is a required parameter for the class to function.