Refactor the getting of Contentful Entries to return an object instead
Before this change the service would query the Contentful API an entry and then call .raw on it to get the full JSON.

This works means consumers of this class have to be aware of the fact it has to deal with a JSON blob, rather than an object.

To avoid `dig`ging everywhere we use `GetContentfulEntry` and to allow us to pass the result to subsequent classes more easily, we stop calling .raw on it straight away. We still call .raw but now only where we set the .raw field on Question.

Not having to stub a JSON file response means we can simplify the test suite slightly so that instead of passing file names around we can pass around a test double.