Refactor the creation of questions to construct the answer
As we include text aswell as radio types the type of answer will start to change. We may want to validate different things on free text answers as opposed to preselected radio answers, for example email address validation.

To open the current code up to this possibility we change up some responsibilities. The controller now gets the entry first and injects it. This isn't strictly required for this code to function but separating these responsibilities helps with testing and would allow us to introduce new services like `CreateRadioQuestion` if we need to become more specific.

Right now we move the construction of an Answer object from the controller into the service. This is a straightforward move at the moment but I am hopeful we this will help us create and use subclasses for form validation. For example: `RadioAnswer`, `ShortTextAnswer`, `LongTextAnswer` etc. We could even go further and have `EmailAnswer` for specific validation.