Short text questions are shown to users for answering
Short text is not the same as long text that might require a larger text
area input. That will come as part of future work.

The fixture includes a real example for a short_text field for an email
address. Any email specific validation can come later if indeed this
question is needed for real. To do that we might start to make different
Answer objects, for example `RadioAnswer < Answer` might include email
validation.