Answers are capitalised on the way in rather than the way out
A previous decision had been made to capitalise answers when rendered on plan#show. This worked but could perhaps be done better on the way into the service. It is fair to assume that the first letter of a radio or text answer should be capitalised.

We currently keep lists of radio options as lowercase in Contentful eg. "catering", "engergy". This may change in future to be "Catering" in which case this line won't strictly be needed but it adds a bit of assurance that answers all start with a capital letter. This is an easy catch to improve the free quality of answers.