Foreign key contraints to ensure data integrity
After this migration if a `plan` or a `question` is deleted then it's associated records will be deleted too, rather than becoming orphaned in the database.

There is no scenario we can think of where a question or an answer should exist without it's parent. Answers only make sense in the context of a question, Questions only make sense in the context of a plan.

Having the database in an expected state where there are no orphans and being able to expect every answer to always have a question provide a helpful level of integrity for us to work with.

We could use `:restrict` if we want to prevent deletions entirely in future. It seems a bit heavy handed for the moment given there is no ability to delete anything.