Refactor Dockerfile so that build-args decide what gems to bundle
Before this change we weren't configuring any parrallel jobs and other config to make it more efficient. Instead of repeating these on both lines, we are now prompted to use the new `config set` syntax by Bundler as a deprecation. This allows us to use a build arg to enable Docker Compose to load different gems on top of the default production gems.