Restarting the server in dev will install gemfile changes
In development we often add, remove or change gems. This requires a bundle before Rails will run.

Before this change the only way to use Docker in dev and deal with Gemfile changes was to completely rebuild the images using `docker-compose up --build --detach`. Even with Docker layer caching this still takes several minutes which is a problem during development.

For images used in production it is desirable to have a static dependency tree. We don't want to be adding extra things at run time as this introduces a chance it could fail because it's not self contained, and may take longer and fail healthchecks.

As a compromise we add a check into the entrypoint which runs at every server start in dev and prod. If it's not on prod, we will see if it needs to be bundled. Given the app directory is mounted into the container the gemfile.lock is correctly updated back into the developers git index for commiting + visibility.

This is not ideal but I can't find a neater solution that satisfies fast dev servers and stable producton images.