Getting Contentful entries goes through a read cache
Whenever we try to lookup a Contentful entry we first check to see if that key already exists in the Redis cache. If it does, we will use it.

The longer term plan is to proactively warm the cache by making requests at 3am. 48 shouldn't get in the way of that but it should cover any unexpected failures with Contentful for a long enough duration to give them a good chance of recovering.

We use a aenvironment variable to allow us (the dev team) to turn this caching off if we need to. This will allow us to disable it on the Contentful preview environment that we have, as well as turn off in emergencies or to force fresh requests.