Fix the way we read JSON from the Redis cache
A string was being passed into the Redis cache that could not be parsed by JSON.parse.

Instead of blindly storing the .raw JSON blob, we now serialise and deserialise it with JSON.dump and JSON.restore.

There is a warning on the use of the JSON.load method [1] however since we are only using it to store trusted content (from Contentful) which only the team can edit, this is a "client under our control".