Rename Plan to Journey
This change creates an abstraction for the concepts of planning and specifying. Functionaly they are both a journey of steps that are defined in Contentful. Defining whether they are a plan or a speification is done by the starting entry_id.

The JourneyController#new action is still connected to the concept of a 'plan'. Now that we're building a specification rather than a plan, this key should probably be renamed. I imagine in future the type of journey "starts" can send a little piece of data to this endpoint to help the action create the right journey.

Questions will be renamed to Steps in a seperate commit.