Remove data migrations from old database migrations
These old migrations now reference a model called `Question` that has since been removed.

We used these database migrations to also migrate existing data within our database, aswell as the schema which is the primary function. We used the ActiveRecord DSL to help us do this by calling `Question.x`.

On our new Heroku Preview applications a fresh database is provisioned and so all the migrations are run. We want to make sure that if the pull request that was being tested included a migration then it would be run in full, so we cannot change the provision instruction from `db:migrate` to `db:schema:load` - even if we had a way.

For data migrations we could look at using the data_migrate gem. We could use rails runners and coordinate running them in between migrations that added a new table and then removed the old table. A far cheaper option is the approach of this commit. After ensuring that this migration has been applied to all environments where real data is concerned we can remove the part of the migration that was concerned with updating the old data across.