Common caching behaviour extracted into reusable service
This approach abstracts the common functionality that we might have of a cache.

This class can be initialised with configuration that can contain  information like whether to enable or disable, and what the key should be. By using a class over a mixin we reduce the amount of parameters required for each method and also make the object injectable, which makes it interchangeable.

I imagine another class `ListEntry` could reuse this Cache and provide it's own configuration to it.

There is a certain amount of logic tied to dealing with a Contentful Entry as the value. Serialising the JSON and creating Contentful Entry objects. Moving this logic to the Cache class would increase it's responsibilities and couple it directly to the caching of entry data only. The logic should probably be moved to a mixin when the ListEntry service looks to add data to the cache in a similar way.