Fix refresh bug that allowed steps to be skipped
Users could arrive on `questions/new` and instead of answering the required question, they could refresh. This would trigger the code to get the next question, and redirect the user to that instead!

The new a ction now creates the question and redirects them to the show action, where no matter how many times they refresh they can't get to the next question UNLESS they manually type the URL back to `question/new`

Since Question#new no longer needs to be responsible for variables needed in the few (now replaced by #show) we don't need an `@answer` returned by the `CreatePlanningQuestion` service. This service is slightly trimmed down to be simpler and the `AnswerFactory` moved up to the controller. This feels like a readable change.