Questions can have a body field
For the Contentful model 'static_content' we will need to expect a body of content to display. We need to store a copy of this and it doesn't make sense to push it into `help_text` which is the only similar field.

Body is optional so as to allow Questions not to have them.