Steps store the Contentful Model
Now that we are handling more than just questions (staticContent too) we would like to ask Steps to behave slightly differently when presenting them, based on which type they are.

In this case static content will never have an answer and it makes no sense to display these back to the user as it wasn't content they added.