Add new class to get all entries from Contentful
This reuses the existing `ContentfulConnector` to call the `entries()` method instead of `entry()`, and return the array of entries.

There is no caching implemented in this class, as we do not yet have a use case where we would want to retrieve all entries from Contentful without also wanting to only use the latest versions.