Rename question constant for readability
The name `ALLOWED_CONTENTFUL_QUESTION_TYPES` where `QUESTION` exists is a problem. Now that the class (now called Step) handles questions and static content, it doesn't make sense for it to include "paragraphs" since they're not a question.

We could either make a new constant and use it in error handling or throw them in together under an abstracted name of `ENTRY_TYPES`.