Sidekiq Cron installed to enable scheduled tasks
Sidekiq can manage and process routine tasks in a similar way to Cron. The advantage to using Sidekiq for this now is that it makes it easier to extend the service with other async tasks. There's a good chance we'll need more of these and to then have to configure and look after a cron _and_ sidekiq process is more costly and introduces another box, which makes the overal architecture a little more complex.