Refactor cache class so a key has to be passed to each method
Setting a key on initialise was perhaps trying too hard to reduce the amount of code needed.

Whilst the approach works when setting up a cache for a single entry, the same design doesn't work when we want to use the same cache and interact with many different entries. We shouldn't have to initialise a new cache object each time there's a new key.

This change changes the approach so that the cache key needs to be passed as a parameter to each method. Less clean but is preferable to modifing a cache object's key data during an enumeration.