Primary form button text is configurable
This is an optional field. If it's not provided or forgotten there's a fallback of "Continue".

Next steps would be to create a StepPresenter for the view logic.

We may need secondary buttons "skip" or tertiary buttons like "cancel" or "back". We use the name "primary" in anticipating for that.