Introduce CheckboxAnswers
This model will collect a response that's an array of strings. One for each option selected and an empty string for one not selected.

If a question Step has options of ["Foo", "Bar"] then the associated `CheckboxAnswers` response could be ["Foo", ""] where "" was unchecked. This array is what's passed along from the form so it seemed the most compatible route to then store this data (rather than converting it to a hash for an hstore) so that when it's used for an edit form, the field is going to be able to pre check the already selected fields. That's the hope but without an edit form it's not simple to test.

I'm using "Answers" in the plural because by the very nature of the checkbox type, multiple options are presented and multiple options can be chosen.