Helper method to construct objects for a checkbox form element
This array is currently identical to the previous `radio_options` method.

I could have renamed that method into `options` to make it reusable for 2 different types of form element. I suspect we may wish to keep the flexiblity to change how we construct these so we can change their behaviour independently.