Validate that at least one Checkbox item is checked
If no checkbox is selected then `[""]` is passed to the controller. This is then stored on the model as such and passes the presence validation. In order to fire the validation when no item is chosen, we only store the chosen options so that if an empty array is validated against it behaves as expected.