Rename test helper for clarity
Step and Entry are names for a very similar thing. Step is the internal represenation of what a Contentful Entry is.

I'd like to create a new method called `fake_contentful_entries` which is the reason for this change. I think it would make sense with this existing method renamed to be the singular version.