Journey map links each step to Contentful for editing
Building upon the listing of entry ids, we now use the title to make them human readable and link them to contentful. The user will be able to use this to reference the exact entry coming from Contentful and make changes.

The environment variables "CONTENTFUL_SPACE" and "CONTENTFUL_ENVIRONMENT" are not secret values that a member of the public could use to access anything. There is perhaps a small bit of security through obscurity to be gained by hiding these but ultimately we are trusting Contentful's authentication mechanism. In future, this page will also be behind an authentication mechanism of our own.

We use target blank for the link as I suspect the user will be jumping between our service and Contentful to compare and review a number of entries.