Some feature tests no longer depend on correct ENV
If CONTENTFUL_PLANNING_START_ENTRY_ID is set locally as a different value, this test starts to break.

We should refactor the way we are setting up all these feature tests by mocking the ENV and handover the majority by stubbing a journey with the desired `journey.next_entry_id`.