Warming the cache only adds new content if valid
If a valid contentful journey is found, we will continue to add those new items to the cache for 24 hours.

If the `BuildJourneyOrder` service raises an error to the effect that the journey was invalid, we instead take the approach of extending all prior entries. This should ensure that a single valid journey remains in the cache for users to use.

Rollbar alerts are managed by `BuildJourneyOrder` so the team should be alerted when this happens.

We do some extra work to make sure we combine TTL times to avoid the case where we actually lower the TTL overall.

Though we only have redis keys/items that are entries at the moment, we make sure we don't accidentally scoop up other types of item by checking the key prefixes.