Access a Step's raw response returns a hash
Rather than a serialised string.

We as a contentful object for its `raw` data. This returns a hash. When we save this against a `binary` field in the database it is serliased into a string. This isn't very useful when we want to use it.

Convert the serialised data into a json serialisable string. Then ask JSON to parse it until we are back where we started with a hash.

A better fix might be to look at changing the field type from `binary` to `jsonb` to see if we this serialising can be done for us.