Every Step has its associated Contentful ID
For existing Steps we tap into each Step's raw attribute to grab the original entry ID, and set it.

As a foreign key that we are going to start depending on when generating a specification this link is important. We set database presence validation to ensure it exists for all records.