Raise an error if the journey is longer than 50 steps
The number 50 is arbitrary. We could have picked 100. This is used as a blunt instrument in order to detect where a journey does not have an end. A single journey where a user has to click through more than 30-40 steps seems enormous and should perhaps be broken down. The number 50 tries to round up to a number we should never realistically hit. Though if we do, it's easy to change.

If there is no end within a map of 50 entries, presume it's incorrectly configured and therefore invalid.

This is not the total count of entries coming back from Contentful, it's the total within a journey.

Given the journey map is not put into the cache, there's no way we can render the last working journey. This may also confuse the user when they're told there was a problem whilst being able to see te last healthy map. We fail loudly for now.

Creating a real fixture with 50 entries brought the test suite to a crawl. It took 2 minutes. For this reason I use a smaller fixture and stub the limit to lower the threshold for when the error is thrown.