Fix migration so presence validation is applied at the end
Applying this straight away fails because of course the new field has no value at all. This shouldn't have got through local testing but alas it did and now we have to fix it.

Create the new temp field without validation, and add it at the end. By adding it at the end, after populating the columns, we should the be told if we still have missing values. That failure is the one we want.