Refactor creation of Journey into a service
We want to start doing some more work when we create a Journey. We're thinking we need to reach out to Contentful to get another type of entry that will tell us - through a liquid blob - how we should render the questions back to the user.

In order to make this extra call and save extra data to the record, we first extract what we had from the controller into a service to help us extend the behaviour with unit test coverage.