Journey stores a copy of the Liquid template
This template will eventually be sourced from Contentful. For now we are faking that with a hardcoded file.

The contents of this file is prone to changing and older versions of it will not guarantee backwards compatibility. For instance an entry_id in the template may no longer exist, or the values it's trying to match on may have been switched out. In order to deal with this versioning issue our general approach is to take a copy of everything in contentful when the user starts a journey. This should ensure it is always able to generate a spec as all the state is self contained.

We use the JSONB type as it's what we use to store the raw response from Contentful entries. Given that this data will eventually come back from Contentful int he same manner. The type could probably be :text: or even :binary: if we wanted it to be.