New service to find a the liquid template
The main reason this service was created was so that we could use it to stub the loading of files in the feature tests and unit tests of CreateJourney. Failing to do this means we have to stub `File.read` which conflicts with other parts of the stack that use this method.

Having a small class that loads this isn't a bad thing. It will give us a natural place to switch the loading of the Liquid template from a hardcoded file to contentful response data.