A user can download their specification as a Word document
In order to construct a document we need to add basic page layout around the specification contents with a <head>, <body> etc.

This mechanism works when the HTML is valid but if it's invalid it will throw all the words into the document without feedback. As an example, if a, <h2> is closed by accident with a </h3> the document loses all it's formatting.

When we use this same HTML as a partial without our ERB show template, it is more forgiving and will _somehow_ make the correction for us. Though there is still an issue if we make a serious omission.

We arguably need to validate the HTML aspect of the Liquid template when we create the journey at the very least. Not something that sits with this commit.