Users cannot start a new journey if the Specification is invalid
We need some certainty that the Specification that we take a copy of is valid as such that it can be used to later render the Specification template.

We cannot raise an error and allow this template to be updated retrospectly since there is no mechanism to update Journey's that are in progress.

Given this we have to fail loudly that the specification is invalid at the point the journey is started. Flagging a message in Rollbar when this occurs to alert the team to the issue. Once this is fixed the user should be able to start the journey.

This is not the only way to protect ourselves from this scenario. Contentful could eventually do it's own validation before new versions of the template are published. We are not there yet though so this is a good first step.