BuildJourneyOrder raises exception on missing entry
Previously, it was possible for entries to refer to a non-existent entry
as part of their `next` field. This change removes the default `nil`
value which would be returned in this case, and replaces it with raising
a new `MissingEntryDetected` exception.

Also updates the fixture for the `WarmEntryCacheJob` spec, which had an
invalid `next`.

Co-authored-by: Lorna Harwood <lorna@dxw.com>