Remove unnecessary test for order of entries in BuildJourneyOrder
This test does not appear to be necessary; it is checking that the order
of entries returned from `BuildJourneyOrder` matches the order of
entries in the fixture, but the order of entries returned from
Contentful is not guaranteed. This test appears to be checking that our
fixture has entries in the order we would expect them to appear, and not
that the application itself puts them in the right order.

Co-authored-by: Lorna Harwood <lorna@dxw.com>