Add a production.liquid template to generate production specs
I had a look to see if we could use the slug (which I have included as a comment to help with discovery) to replace the entry_id, and in doing so the need for different liquid templates. Given that the slug field is editable it is not stable enough to depend on.