Fix test so that step.raw returns as a hash rather than string
We expect the raw method to return a `hash`. This is expectation ensures we can reliably interogate it with methods such as `fetch`.

I found that a couple of tests [1] that don't use `fake_contentful_entry` but instead only rely on `stub_get_contentful_entry` to stub out a fake Contentful response. I found that the behaviour is different for how raw is stubbed. In this method we set `raw` to the string contents of the file, but in `fake_contentful_entry` we have an extra step to parse the string into JSON.

This change removes the redundant code in `stub_get_contentful_entry`. We can see that only 1 test actually fails. We change that failure to check the exact output we expectm, instead of trying to be clean and use parsing methods in the test. This no doubt patched over the problem and made it harder to spot.

[1] ./spec/features/visitors/anyone_can_complete_a_journey_spec.rb:29