Radio form elements support the new JSON format
We are introducing a new `extendedOptions` field which will eventually replace the `options` field completely.

This change converts any entries/steps that use the old options field into a hash which the radio form element can interpret and render. Right now the behaviour remains exactly the same. With a hash of options now available we open the door to being able to use different configuration in this question such as additional help text or supporting a call for further information [1]

I've left several TODO comments as I'll need to come back as part of this work and tidy up once this change is deployed and all data is migrated across.

Select box pattern uses the same options field but doesn't require a hash of options yet. This work will make that possible but doesn't jump to changing the views before we need to.

[1] https://govuk-form-builder.netlify.app/form-elements/radios/#radio-buttons-fieldset-with-conditional-content-and-a-divider