Fix test suite so that factories have the correct type
Whilst we have feature coverage that tests this doesn't affect users this does make our test suite inaccurate.

When we manually create or fake a `single_date_answer` of `checkbox_answers` they don't have the `contentful_model`. This is used in the journey#show view to only replay steps that were questions back to the user, as opposed to static content.