Fix checkbox answers can be edited
This change follows the creation pattern of checking the model before grabbing the right parameters for assigning. We can't just set assign with a single `params` variable when an included parameter is an array.

The factory needed to be changed slightly with an additional `options` field. This makes sure that when the attached answer is edited, the same list of options are presented back to the user. Again this doesn't change behaviour it only makes our test suite more realistic.