Docker Compose server start auto purges any pid file
When we start the webserver it creates a pid file in tmp/.

When running the server locally with docker compose and you turn off your machine without gracefully terminating the server process the file remains.

When you start the server again you get an error asking you to remove the existing pid file.

This change automates this step so we don't have to do this manual step each time.

It doesn't affect production environments.