Users can be asked Radio questions that prompt for additional text
Use this [1] pattern from the DFE Form builder gem. We take the behaviour that is applied to the 'Other' option in the example and make it applicable to any option.

Contentful now always supplies JSON for a list of options. It is now possible to provide more than a simple list of values. If the Content user provides keys "display_further_information" and "further_information_help_text" then the question will render an extra text input. This is per option and is not required.

We think we may want to toggle validation through configuration in future. For now these text fields are optional and won't fire validation errors.

[1] https://govuk-form-builder.netlify.app/form-elements/radios/#radio-buttons-fieldset-with-conditional-content-and-a-divider