All steps in a journey are now loaded at Journey creation
To support the task list pattern showing unanswered questions, when a
user starts a `Journey` we now copy all steps in that journey from
Contentful and create the necessary database entries.

As a result of this, we no longer load a step at any point after the
creation of a `Journey` (eg when visiting the 'next step' action), which
has meant several similar changes to tests. Tests for these features now
use the pattern of starting a `Journey` which covers a complete
multi-question journey, and then selecting the correct question page to
visit directly.

Fixtures for these single-case features have been deleted. Some remain
which are to support tests on present-but-unreachable paths, which will
be removed in a later tidying-up commit.

Co-authored-by: Lorna Harwood <lorna@dxw.com>