Responses used in Specifications are formatted by presenters
This fixes an immediate issue we can see where radio answers show "Yes" without their `futher_information`. It should be "Yes - Further information" and the logic already exists within presenters.

We need to get a fix for this quickly to unblock the production deploy. I've left a note to prompt us to refactor this controller code the next time we pass it.