Previewing a question now specifies a liquid_template
`liquid_template` is a required part of a `Journey` which was not being
provided whilst previewing a question, leading to the `Journey` not
being persisted and a redirect to the step within the journey being
impossible.

This commit adds a very short stub template to preview journeys.

It also adds a feature spec to ensure that a user attempting to preview
a step does reach the intended view.