Only display specification if all tasks are completed
The new all_tasks_completed? method within the Journey class tests to
see if all Steps within the Journey have a matching Answer. If it comes
across any Step which does not have an answer, it short circuits and
returns false.

This method is then used to selectively display the specification and
spec download options in the task list only if all tasks have been
completed.

Co-authored-by: Lorna <lorna@dxw.com>