Refactor how we get a Category so it can handle liquid template validation
Change the behaviour so that the getting of a category is not directly tied to the getting of category steps. Instead we find the category first (using the same GetEntry service under the hood) and then inject that onwards into the GetCategoryEntries service.

This change is being made at this poiint because we have a new need. The Category Entry will begin returning a new field called `specification_template`. This value needs additional validation to make sure it looks like a value that Liquid can parse. We should not move onto loading steps from a category unless the category itself is valid. Moving this into a 2 step process therefore makes sense and gives us a class to add further validations in future.

Testing improvements can also be made where we can now inject a category object into tests instead of stubbing additional calls. We make some tweaks to how many times we expect the ContentfulConnector so it is an accurate representation of the real world. When we ask for a category, we only expect 1 call. If we have a category with steps we expect exactly `steps.count` number of calls to be made.