Remove expectations on connector initialisations
When trying to introduce 'sections' into the mix I got heavily bogged down by trying to ensure this test continued to pass at all stages.

On reflection I don't think the assertion is valuable. The most important thing is that an expected set of objects are returned, it doesn't really matter how many connections were made. I suppose the value this was meant to provide was to prevent making more calls that was necessary. Given we are making 1 request per entry we know we'll be making a lot of calls. I think we can set this back to a simpler stub.